+++ src/config/gro.config.default.ts
@@ -36,8 +36,6 @@ export const config: GroConfigCreator = async ({fs, dev}) => {
 		logLevel: ENV_LOG_LEVEL ?? LogLevel.Trace,
 		types: !dev && enableNodeLibrary,
 		plugin: async () => [
-			// TODO dev server?
-			// enableDevServer ? (await import('../plugin/groPluginDevServer.js')).createPlugin() : null,
 			// TODO some usecases may need to run the API server during the build for e.g. prerendering,
 			// but it's currently disabled because the adapter-node usecase has the production API server
 			// depend on the middleware created later in the adapt step of the build