+++ src/lib/app/FileTreeExplorerNode.svelte
@@ -1,7 +1,0 @@
-<script lang="ts">
-	import type {FileTreeNode} from '$lib/app/fileTree';
-
-	export let node: FileTreeNode;
-</script>
-
-<div class="node">{node.name}</div>