+++ src/lib/app/SourceTreeVisualizer.svelte
@@ -1,6 +1,0 @@
-<script lang="ts">
-	export let name: string;
-</script>
-
-source tree visualizer:
-{name}