+++ changelog.md
@@ -1,9 +1,11 @@
 # changelog
 
-## 0.54.1
+## 0.55.0
 
-- always build before running tasks
+- **break**: remove the dev server, including `gro serve` and project metadata
   ([#321](https://github.com/feltcoop/gro/pull/321))
+- always build before running tasks
+  ([#322](https://github.com/feltcoop/gro/pull/322))
 
 ## 0.54.0